Moved GeoTrellisMethods to its own file, GeoTrellisMethodsWrapper
